refactor(llm-agent): move ToolCache and IToolCache from llm-agent-server
Move cache-related types and implementations (IToolCache, ToolCache, NoopToolCache) from llm-agent-server to llm-agent. Server files become re-export shims pointing to the new location in llm-agent's public API (Task 11 wires the exports).

Changes:
- Create packages/llm-agent/src/cache/ with moved implementations
- Move test file packages/llm-agent/src/cache/__tests__/tool-cache.test.ts
- Replace server cache files with re-export shims
- All 218 tests pass, including 9 cache-specific tests
