+++ packages/llm-agent-server/src/smart-agent/cache/index.ts
@@ -1,3 +1,2 @@
-export { NoopToolCache } from './noop-tool-cache.js';
-export { ToolCache } from './tool-cache.js';
-export type { IToolCache } from './types.js';
+export type { IToolCache } from '@mcp-abap-adt/llm-agent';
+export { NoopToolCache, ToolCache } from '@mcp-abap-adt/llm-agent';
