+++ packages/llm-agent-server/src/smart-agent/cache/noop-tool-cache.ts
@@ -1,9 +1,1 @@
-import type { IToolCache } from './types.js';
-
-export class NoopToolCache implements IToolCache {
-  get(): undefined {
-    return undefined;
-  }
-  set(): void {}
-  clear(): void {}
-}
+export { NoopToolCache } from '@mcp-abap-adt/llm-agent';
