+++ packages/llm-agent-server/src/smart-agent/cache/types.ts
@@ -1,14 +1,1 @@
-import type { McpToolResult } from '@mcp-abap-adt/llm-agent';
-
-export interface IToolCache {
-  get(
-    toolName: string,
-    args: Record<string, unknown>,
-  ): McpToolResult | undefined;
-  set(
-    toolName: string,
-    args: Record<string, unknown>,
-    result: McpToolResult,
-  ): void;
-  clear(): void;
-}
+export type { IToolCache } from '@mcp-abap-adt/llm-agent';
