+++ packages/llm-agent/src/cache/noop-tool-cache.ts
@@ -1,0 +1,9 @@
+import type { IToolCache } from './types.js';
+
+export class NoopToolCache implements IToolCache {
+  get(): undefined {
+    return undefined;
+  }
+  set(): void {}
+  clear(): void {}
+}
